refactor: Add function `calculateNLv` to calulate the nitrogen mineralization from soil organic mattter. Rename `calculateNlvSupplyBySom` to `calculateNlvSupplyIncreaseBySomPotential`
