+++ .changeset/swift-ears-argue.md
@@ -1,0 +1,5 @@
+---
+"@nmi-agro/fdm-calculator": minor
+---
+
+Add function `calculateNLv` to calulate the nitrogen mineralization from soil organic mattter. Rename `calculateNlvSupplyBySom` to `calculateNlvSupplyIncreaseBySomPotential`
