+++ fdm-calculator/src/index.ts
@@ -139,8 +139,8 @@ export type {
 export { collectInputForSoilParameterEstimates } from "./estimates/input"
 export { getSoilParameterEstimates, requestSoilParameterEstimates } from "./estimates/index"
 export type { SoilParameterEstimatesInput, SoilParameterEstimatesResponse } from "./estimates/types"
-export type { NlvSupplyBySomParams } from "./other/nlv-supply-by-som"
-export { calculateNlvSupplyBySom } from "./other/nlv-supply-by-som"
+export type { NlvParams, NlvSupplyBySomParams } from "./other/nlv-supply"
+export { calculateNlv, calculateNlvSupplyIncreaseBySomPotential } from "./other/nlv-supply"
 export type { WaterSupplyBySomParams } from "./other/water-supply-by-som"
 export { calculateWaterSupplyBySom } from "./other/water-supply-by-som"
 export type { CultivationForHoofdteelt } from "./shared/hoofdteelt"
