+++ fdm-docs/blog/2026-03-04-release-notes.mdx
@@ -68,7 +68,7 @@ This release introduces a formal invitation system for granting access to farms,
 
 Two new functions have been added to quantify the agronomic value of changes in soil organic matter (SOM):
 
-- **`calculateNlvSupplyBySom`** — calculates the change in nitrogen-leverend vermogen (NLV) resulting from a given change in SOM content.
+- **`calculateNlvSupplyIncreaseBySomPotential`** — calculates the change in nitrogen-leverend vermogen (NLV) resulting from a given change in SOM content.
 - **`calculateWaterSupplyBySom`** — calculates the change in topsoil water holding capacity resulting from a given change in SOM content.
 
 These functions complement the existing organic matter balance capabilities and can be used to evaluate the practical impact of SOM management decisions on nitrogen supply and water retention.
